third_party/blink/web_tests/external/wpt/css/css-anchor-position/scrollable-containing-block-validity-expected.txt

This is a testharness.js-based test.
[FAIL] .target 2
  assert_equals: \n<div class="target" style="position-anchor: --a;" data-expected-width="200" data-expected-height="200"></div>\nwidth expected 200 but got 100
Harness: the test ran to completion.
